second commit update employee

+++ src/main/java/com/neo/sihotel/repository/TypeRoomRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 import javax.persistence.criteria.CriteriaBuilder;
 
 @Repository
-public interface TypeRoomRepository extends CrudRepository<TypeRoom, Integer> {
+public interface TypeRoomRepository extends CrudRepository<TypeRoom, Long> {
 }
